+++ library/core/src/main/java/com/google/android/exoplayer2/upstream/BandwidthMeter.java
@@ -15,6 +15,8 @@
  */
 package com.google.android.exoplayer2.upstream;
 
+import android.support.annotation.Nullable;
+
 /**
  * Provides estimates of the currently available bandwidth.
  */
@@ -40,4 +42,10 @@ public interface BandwidthMeter {
 
   /** Returns the estimated bandwidth in bits/sec. */
   long getBitrateEstimate();
+
+  /**
+   * Returns the {@link TransferListener} that this instance uses to gather bandwidth information
+   * from data transfers. May be null, if no transfer listener is used.
+   */
+  @Nullable TransferListener<?> getTransferListener();
 }